Yet another late welcome for git.
I've done the entire backend for this, as far as I can tell so far, so now it's
just a matter of creating the interface in curses. All the db stuff, keyword
parsing, command line parameters, rc file parsing etc. is done so any changes
to them now will be logged in git, otherwise it's interface all the way from
here no out.