Extreme interface work done
I should probably have done a bunch more commits on this but I was on a roll.
The interface just went from zero to something that actually works as an
interface - two panels, one on the left with a list of notes in it, one on the
right that displays the note itself. Selecting up and down works (with the
usual highlighting of the current note) and keywords are highlighted in the
body of the note. All of the colour code parsing and everything like that was
all done today too. So a bunch of things was done. Hopefully now I can do some
more refined commits. :-)