Playing with touchwin/redrawwin to recreate segfault
amk didn't seem too convinced that this was a problem with pycurses.
Well now I have two use cases that show that it basically _is_ so
this commit was just me playing with that. I've added back the touchwin
bit on hard_update() but it never seemed to do much unless redrawwin
was called as well. Unfortunately redrawwin makes Python segfault. Wham.